feat: update TaskCard buttons to improve UI/UX
- Replace text Complete button with CheckSquare icon for a cleaner UI
- Replace Info icon with a circular "i" button for better visual recognition
- Reorganize buttons layout with Complete action on the left and other actions on the right
- Ensure consistent button styling across both equipped and repository modes
- Hide Complete button for already completed tasks

+++ server/controllers/cardController.js
@@ -295,7 +295,7 @@ const issueBlankCard = asyncHandler(async (req, res) => {
     type: "blank",
     title,
     description,
-    taskDuration: "short", //  限定为short卡片
+    taskDuration: "long", //  限定为short卡片
     issuedAt: new Date(),
   });
